Optimisations
When sending SPI data, MOSI state is now only set when changed.
MOSI state is now always returned to low, at the end of an SPI send
(could have caused a defect, if power-down bit PD0 was set high.)
When receiving SPI data, only the exact number of expected bits (8 or
12) are recevied. Zero padding is skipped.
Together this yeilds approximatley a 15% speed improvement when reading
the channels.